autoload/auditory.vim: Enable scrolling mappings with FIXMEs
This way the sounds on scoll commands acutally work. They no longer
support counts :( and the <c-b> command for some reason needs to be
entered twice in order for it to work. Enable these for now just so we
can get the sounds but add FIXMEs so we know that these are borked.