plugin/auditory.vim: Show warning if mplayer not installed
On plugin load, if mplayer isn't installed then fail ungracefully. Since
we depend on mplayer in order to play any sounds, my thinking now is
that I want to fail immediately.

Still trying to decide if this is be best course of action. Maybe I just
want an informative message instead of a complete failure.