plugin/auditory.vim: Decrease warning level of mplayer error
Instead of echoing and error, just echo a message about mplayer not
being installed. Errors look pretty annoying on initial load, and I
don't think this is worth the red background colour. It could be that a
user has two machines, one with mplayer installed and one without. In
that case, I don't want people to have to disable the plugin on one
machine. That seems like too much to ask.

I think this is a nice middle-ground between alerting users that
something is needed by the plugin and not barfing if it's not there.