autoload/auditory.vim: Delete auditory#NormalModeMappings()
Now that we're assigning our mappings using the mapping dictionary, we
can remove the old function that applied the mappings manually.

It definitely less clear an readable I'll admit, but this allows us more
flexibility and hopefully customisability. The original plan with the
mapping dictionary was to allow toggling of the mappings on and off, so
hopefully it will help with that.