autoload/auditory.vim: Only map with `<silent>` if using `execute`
We don't need to use `<silent>` in our mappings if we're going back to a
regular non-command mode command, so listen for commands where we
`execute` in the mapping and don't `<silent>` them.

Reverted the ternary I had added previously to make this workable with
the now 2 variables that need to listen for `execute`.

Also add some line wrapping to the map creation for better readability.