autoload/autoload.vim: Use `nnoremap` instead of `nmap`
I wanted to use `nmap` so that user-defined commands would come through,
but that caused the sounds to play recursively, a problem I now remember
from when I originally wrote this. Change it to `nnoremap` by default to
get around this. We'll look into whether it's possible to get
user-defined mappings attached somehow. Maybe in the save-and-restore
process.