autoload/auditory.vim: Ensure we `<cr>` if we're not piping
If we're not piping to an `execute` command in the mapping, we need to
`<cr>` to execute the sound player and type the normal mode command.
Otherwise the normal mode command gets typed in command mode which is
not what we want and messes things up greatly.

Figured I would just tack the `<cr>` on to the pipe variable rather than
create a new one since it's the inverse of what we need in that
situation. Hopefully the intention is clear enough but I'll grant that
this doesn't expose the most clarity.

Edit: Added a comment for better explanation of the <cr>.