autoload/auditory.vim: Unsilence search commands
The search commands, as I has specified previously in the verbose
mappings from before, should not be silenced, otherwise you can't see
the useful output that they provide in the command line area.

Add a way to make a mapping silent or not using a new field in the
mappings dictionary.