autoload/auditory.vim: Restore user mappings when turning off
In `auditory#Unmap()`, after unmapping Auditory maps, remap user-defined
mappings.

Say you redefined `j` -> `gj`. When turning Auditory off, we would
previously just unmap `j`, resulting in `j` -> `j`. Now that we're
restoring user mapppings, `j` once again gets mapped to `gj`.

Also keep track of the mode that Auditory uses for its mappings when
storing a user-defined mapping so that we're only saving and restoring
them if they're defined in the mode used by Auditory.

User mappings get saved in `auditory#AssignMappings()`, or the function
that essentially turns on the plugin.

Fix the `if` condition that checks `if l:user_mapping`. This gave me a
lot of trouble because I needed to match it against emtpy string,
instead of a boolean condition. Use the proper condition to get it
actually working.