autoload/auditory.vim: Move `d` mappings to `s:mappings`
Put our `d`, `v_d`, and `dd` mappings in the `s:mappings` dictionary so
that they get automatically activated and deactivated when our map and
unmap functions go through the dict.

This required some modifications to `auditory#AssignMappings()` to allow
it to accept `map_to` commands that don't have a preceding
`auditory#Play()` call (because in the case of our `d` mappings, the
audio is played inside the function they call.

Also ensure we're not adding a pipe to the command if there's no
preceding audio part.