autoload/auditory.vim: Fix `AuditoryToggle` + `Auditory[On|Off]`
When using `:AuditoryToggle` in conjunction with `:AuditoryOn` or
`:AuditoryOff`, things would get messed up. This is because
`:AuditoryToggle` relies on `g:auditory_on` in order to know whether to
turn sounds on or off, but the variable was only getting set in
`auditory#ToggleMappings`, not in the functions that `:AuditoryOn` or
`:AuditoryOff` call.

Move the setting of the variable into `auditory#AssignMappings` and
`auditory#Unmap` so that the right thing will happen regardless of which
on/off/toggle commands you run.