Don't turn on Galaxy Far Far away when not `g:auditory_on`
My change from 71ec214cdb8814ecb71c837bf947583fda54d37e caused Galaxy
Far Far Away sounds to be turned on any time you had
`g:auditory_galaxy_far_far_away` turned on, regardless of whether
`g:auditory_on` was also turned on.

Fix this because we never want sounds to play when `g:auditory_on` is
off.

We now rely on `auditory#AssignInsertMappings()` to know whether to use
Galaxy Far Far Away or default insert mode sounds.

Need to toggle `g:auditory_galaxy_far_far_away` in that function because
we're calling the toggle function so we need to set it to the opposite
of what we want.

In the future if we add more insert mode sounds we'll have to change
this structure around but it should work for now.