autoload/auditory.vim: Remove old functions
Delete some old code that isn't used any more. This was when I first
made the plugin. These functions were used to start a song when insert
mode was activated and stop the song on insert leave.

Since we're not playing full songs any more, and our insert sounds are
tied to key presses, these functions are no longer needed.