Fix normal mode `G` mapping
Pressing `G` would bring the cursor to the first line in the buffer
instead of the last line in the buffer because the count in the mapping
was always `v:count1`. This meant that the `G` mapping behaved like `1G`
when no count was specified.

This change fixes the `G` mapping to use `v:count` so that it goes to
the last line in the buffer as expected.

We also add support for arbitrary mappings to choose whether they want
to be `v:count` or `v:count1` depending on what value is specified in
the mapping dict.