autoload/auditory.vim: Prevent double `AuditoryOn` & `AuditoryOff`
Previously executing `:AuditoryOn` and `:AuditoryOff` would always do
the thing they're supposed to do regardless of whether Auditory was
already on or off.

This caused errors because things got messed up when trying to
double assign the mappings and double unmap the mappings.

Now check for the status of the `g:auditory_on` flag before turning
Auditory on or off.